added scripts for resampling, tts and transcoding

+++ compute_grid.py
@@ -16,7 +16,7 @@ def extract_lf0(directory):
 
 
 def extract_mcep(directory):
-    if len(list(glob.glob(os.path.join(directory, '*.mgc')))) > 0:
+    if len(glob.glob(os.path.join(directory, '16k_*.mgc'))) > 0:
         return
     for fn in glob.glob(os.path.join(directory, '16k_*.wav')):
         base = '.'.join(fn.split('.')[:-1])
@@ -34,15 +34,18 @@ if __name__ == '__main__':
     dirs = [d for d in os.listdir(wd) if os.path.isdir(os.path.join(wd, d))]
     res_fn = 'mcd.out'
     for pair in itertools.combinations(dirs, 2):
+        if not (pair[0] == "svox" or pair[1] == "svox"):
+            continue
         print('Processing "{}" vs "{}"'.format(pair[0], pair[1]))
         # extract_lf0(os.path.join(wd, pair[0]))
-        # extract_lf0(os.path.join(wd, pair[0]))
-        extract_mcep(os.path.join(wd, pair[1]))
+        # extract_lf0(os.path.join(wd, pair[1]))
+        extract_mcep(os.path.join(wd, pair[0]))
         extract_mcep(os.path.join(wd, pair[1]))
         if os.path.exists(res_fn):
             os.remove(res_fn)
         # cmd = ['./compute_lf0se.sh', os.path.join(wd, pair[0]), os.path.join(wd, pair[1])]
         cmd = ['./compute_mcd.sh', os.path.join(wd, pair[0]), os.path.join(wd, pair[1])]
+        print(cmd)
         subprocess.call(cmd)
         # shutil.copyfile(res_fn, os.path.join(wd, '{}_vs_{}.lf0'.format(pair[0], pair[1])))
         shutil.copyfile(res_fn, os.path.join(wd, '{}_vs_{}.mgc'.format(pair[0], pair[1])))
